refactor: consolidate admin hooks under a common module and update imports
