+++ src/app/admin/api-access/components/AdminApiAccessPanel.tsx
@@ -1,10 +1,10 @@
 import { useState } from 'react'
-import { useAdminTable } from '@/app/admin/hooks'
 import { AdminPageLayout, AdminSearchFilters, AdminStatsDisplay } from '@/components/admin'
 import { Button, Card, ColumnVisibilityControl, useConfirmDialog } from '@/components/ui'
 import { POLLING_INTERVALS } from '@/data/constants'
 import storageKeys from '@/data/storageKeys'
 import { useColumnVisibility } from '@/hooks'
+import { useAdminTable } from '@/hooks/admin'
 import { type ColumnDefinition } from '@/hooks/useColumnVisibility'
 import { api } from '@/lib/api'
 import toast from '@/lib/toast'
