+++ src/app/admin/api-access/components/DeveloperKeyTable.tsx
@@ -1,4 +1,3 @@
-import { type UseAdminTableReturn } from '@/app/admin/hooks/useAdminTable'
 import { AdminTableContainer, AdminTableNoResults } from '@/components/admin'
 import {
   Badge,
@@ -8,6 +7,7 @@ import {
   RefreshButton,
   SortableHeader,
 } from '@/components/ui'
+import { type UseAdminTableReturn } from '@/hooks/admin'
 import { type UseColumnVisibilityReturn } from '@/hooks/useColumnVisibility'
 import { cn } from '@/lib/utils'
 import { type ApiKeySortField } from '@/schemas/apiAccess'
