+++ src/hooks/admin/index.ts
@@ -1,2 +1,3 @@
+export * from './useAdminFilters'
 export * from './useAdminTable'
 export * from './useReviewRiskFilter'
